refactor: cleanup dev resolver
